+++ app/src/main/java/com/coppel/rhconecta/dev/data/release/model/get_releases_previews/GetReleasesPreviewsRequest.java
@@ -1,6 +1,6 @@
 package com.coppel.rhconecta.dev.data.release.model.get_releases_previews;
 
-/* */
+/** */
 public class GetReleasesPreviewsRequest {
 
     /* */
@@ -8,12 +8,10 @@ public class GetReleasesPreviewsRequest {
     /* */
     public int clv_opcion;
     /* */
-    public long opc_acceso;
+    public Integer opc_acceso;
 
-    /**
-     *
-     */
-    public GetReleasesPreviewsRequest(Long num_empleado, int clv_opcion, long opc_acceso) {
+    /** */
+    public GetReleasesPreviewsRequest(Long num_empleado, int clv_opcion, Integer opc_acceso) {
         this.num_empleado = num_empleado;
         this.clv_opcion = clv_opcion;
         this.opc_acceso = opc_acceso;
