+++ app/src/main/java/com/coppel/rhconecta/dev/data/visionary/model/get_visionaries_previews/GetVisionariesPreviewsRequest.java
@@ -8,12 +8,12 @@ public class GetVisionariesPreviewsRequest {
     /* */
     public int clv_opcion;
     /* */
-    public long opc_acceso;
+    public Integer opc_acceso;
 
     /**
      *
      */
-    public GetVisionariesPreviewsRequest(Long num_empleado, int clv_opcion, long opc_acceso) {
+    public GetVisionariesPreviewsRequest(Long num_empleado, int clv_opcion, Integer opc_acceso) {
         this.num_empleado = num_empleado;
         this.clv_opcion = clv_opcion;
         this.opc_acceso = opc_acceso;
